+++ app/src/main/java/com/fintern/ourbudgeting/data/repository/AssetAdditionRepository.kt
@@ -1,4 +1,4 @@
-package com.fintern.ourbudgeting.ui.assetmanagement.assettypeaddition
+package com.fintern.ourbudgeting.data.repository
 
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FieldValue
